# swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'User API'
paths:
  /user:
    get:
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
      x-swagger-router-controller: user
      operationId: listUsers 
    post:
      parameters:
      - name: body
        in: body
        required: true
        schema:
          type: object
          properties:
            name:
              type: string
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
      x-swagger-router-controller: user
      operationId: createUser
  /user/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
      x-swagger-router-controller: user
      operationId: getUser
    put:
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      - name: body
        in: body
        required: true
        schema:
          type: object
          properties:
            name:
              type: string
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
      x-swagger-router-controller: user
      operationId: updateUser  
    delete:
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
      x-swagger-router-controller: user
      operationId: deleteUser     
 
       

      